New design for structured mesh representation.
Key changes from the previous representation:
1. The old index-space for structured meshes was designed
to contain a single bounding box leading to multiple
index-spaces corresponding to an intermediate entity
such as edges in 2D. The new design removes this constraint
to support multiple bounding boxes and thus provide a
single index-space for entities of each dimension in the mesh.
2. Specialized iterators are provided to traverse the
mesh both locally and globally.